packages/pyscaf-core/pyproject.toml
[project]
name = "pyscaf-core"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "tomlkit>=0.12",
    "click>=8.0",
    "questionary>=2.0",
    "rich>=13.0",
]

[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pyscaf_core"]
